Extracts the StatCounter to a separate file and adds datacontract
